Add /guardian/[studentId] insight page against the 5 live Phase 1 RPCs
Single-screen layout per design doc Module B mockup: header (child's
name via guardian_get_students + Qmon via guardian_get_qmon_display),
7-box weekly calendar (guardian_get_weekly_calendar, today/future
styled distinctly), goal progress (guardian_get_goal_progress, bucket
text rendered as-is, empty result shown as "ยังไม่ได้ตั้งเป้าหมาย"),
top categories with a client-side "ดูทั้งหมด" toggle over the full
list guardian_get_categories already returns (no second RPC call), and
two bottom action buttons linking to /guardian/goal and /guardian/plan
(neither built yet, so they 404 for now).

Skipped the "สิ่งที่คุณส่งไปได้ผล" section entirely — confirmed via the
Notion design doc (Guardrail + โมเดลธุรกิจ + Pilot page) that it's still
an unbuilt idea with no backing table, not something this round missed.

Reuses getGuardianAccess() unchanged for the account-level double-gate.
Per-student authorization relies on the RPCs' own guardian_links checks:
any RPC error (wrong/unlinked studentId) or an empty guardian_get_students
match renders a 404 instead of a broken partial page.

Also synced guardian_get_qmon_display() in the migration file to match
the ::integer cast on pets.stage that was hotfixed directly on
production during testing (pets.stage is smallint) — no other RPC
changed on prod, confirmed against live prosrc.

No new migration needed this round — frontend only, as scoped.

+++ supabase/migrations/20260914130000_guardian_insight_phase1.sql
@@ -318,6 +318,8 @@ comment on function public.guardian_get_categories(uuid) is
 -- ============================================================
 -- 6) guardian_get_qmon_display — Qmon ที่ active ของนักเรียนคนนี้ ทรงเดียวกับ PetDisplayInput
 --    (src/components/social/petSummary.ts) ให้ client เรียก resolvePetDisplay() ต่อได้เลย
+--    pets.stage เป็น smallint จริง (ไม่ใช่ integer) — cast ::integer ให้ตรงกับ RETURNS TABLE
+--    ที่ประกาศไว้ (พบระหว่างเทสจริงกับ ซันซัน↔pond แก้แล้ว live บน prod ไฟล์นี้ sync ตามให้ตรง)
 -- ============================================================
 create or replace function public.guardian_get_qmon_display(p_student_id uuid)
 returns table (
@@ -351,7 +353,7 @@ begin
   end if;
 
   return query
-  select p.nickname, p.stage, p.subline, p.personality, e.sprite_prefix, e.name_th
+  select p.nickname, p.stage::integer, p.subline, p.personality, e.sprite_prefix, e.name_th
   from public.pets p
   join public.egg_types e on e.id = p.egg_type_id
   where p.user_id = p_student_id and p.is_active = true
